requirements.txt
PySide6
librosa
moviepy
numpy
Pillow
proglog
requests